Add validation for read-route.
