+++ src/main/scala/com/github/scalalab3/logs/http/ReadServiceRoute.scala
@@ -10,8 +10,10 @@ trait ReadServiceRoute extends AbstractHttpService {
 
   val readRoute = get {
     (pathSingleSlash & parameter('page.as[Int] ? 1) & parameter('per_page.as[Int] ? 100)) { (num, size) =>
-      complete {
-        pass(readService ? Page(num, size))
+      validate(num > 0 && size > 0, "The number and size of the page must be greater than 0.") {
+        complete {
+          pass(readService ? Page(num, size))
+        }
       }
     }
   }
